+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
-## [v2.2.17.dev0]
+## [v2.3.1.dev1]
+
+### Added
+- [knative] Include GCP and Azure storage libs into default knative runtime
+
+### Changed
+- [IBM CF] Change user_key to API-key pass instead of user 
+
+### Fixes
+- [Cli] Fixed --config param in lithops cli
+- [Standalone] Fixed internal executions
+- [Core] Fixed rabbitmq monitor when get_result() is called after wait()
+- [GCP Storage] Fix GCP Storage backend put obj as stream
+- [GCP Functions] Improved runtime create time
+- [Azure blob] Fix in azure blob get method
+
+
+## [v2.3.0]
 
 ### Added
 - [Core] Added multiprocessing support in workers
@@ -24,7 +41,7 @@
 - [Core] Renamed partitioner 'chunk_n' param to 'obj_chunk_number'
 - [GCP Cloud Run] Refactor backend, removed 'gcloud' CLI calls.
 - [IBM VPC] Improved IBM VPC backend
-- [Aws Lambda] Lambda layer modules update
+- [AWS Lambda] Lambda layer modules update
 
 ### Fixes
 - [Multiprocessing] Fix issues related to Pipes and Queues